Remove obsolete code.
