+++ src/kit/model/index.js
@@ -9,7 +9,6 @@ export { default as ChildModel } from './ChildModel'
 export { default as CollectionModel } from './CollectionModel'
 export { default as createNodePropertyModels } from './createNodePropertyModels'
 export { default as createValueModel } from './createValueModel'
-export { default as FlowContentModel } from './FlowContentModel'
 export { default as JSONConverterNew } from './JSONConverterNew'
 export { default as ManyRelationshipModel } from './ManyRelationshipModel'
 export { default as ObjectModel } from './ObjectModel'
